+++ snowddl/resolver/clone_table.py
@@ -82,12 +82,8 @@ class CloneTableResolver(AbstractResolver):
         )
 
         for r in cur:
-            # Skip external tables
-            if r["is_external"] == "Y":
-                continue
-
-            # Skip event tables
-            if r["is_event"] == "Y":
+            # Skip other table types
+            if r["is_external"] == "Y" or r["is_event"] == "Y" or r["is_hybrid"] == "Y" or r["is_iceberg"] == "Y":
                 continue
 
             tables_for_clone[f"{self.config.env_prefix}{r['database_name']}.{r['schema_name']}.{r['name']}"] = {